Add UI module to ls_dcm project
This commit includes the creation of two new files in the ls_dcm_ui directory: a build script and an 'app.slint' file representing the AppWindow UI component. At this stage the UI is only a demo UI to learn the concepts of Slint. Additionally, it adjusts the Cargo.toml file to include the newly added 'ls_dcm_ui' module in the workspace.