Bug fix: pixel mask is now set using the high bit.
The pixel data mask was originally set using the bits stored property
but this was incorrect. It is now set with the high bit property as is
provided in a DICOM dataset.
Log crate was added as a dependency and the env_logger was added as a
dev-dependency..
Version of the dicom-rs dependency crates were updated.