18 add docs for dcm cp [closes #18] (#19)
* Rename `dcmcp` package to `dcm_cp`

Updated all instances of the `dcmcp` package to `dcm_cp` to improve consistency and clarity. The changes include source files, configuration files, and the package name in the Cargo.toml configuration.

* Added documentation in README.md in `dcm_cp` and the project.

Usage instructions and build commands for `dcm_cp` have been added in the new section of `dcm_cp/README.md`. A link to the `dcm_cp` project has been added to the project README.md

* Changed title of project in `dcm_cp` README.md.

* Added help command under Usage in `dcm_cp` README.md.